flask_app/requirements.txt
requests
flask
gunicorn
rpi.gpio
flask-restplus
rq
rq_dashboard
redis